requirements.txt
certifi==2020.12.5
click==7.1.2
cloudinary==1.25.0
Flask==1.1.2
Flask-PyMongo==2.3.0
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pymongo==3.11.3
python-decouple==3.4
six==1.15.0
urllib3==1.26.4
validate-email==1.3
Werkzeug==1.0.1
